requirements.txt
fastapi
uvicorn
pandas
sentence-transformers
faiss-cpu
pytest
requests
mlflow
streamlit